atulizando questão do campeonato

+++ simulado-segunda-prova/campeonato.cpp
@@ -4,11 +4,11 @@ using namespace std;
 
 int main(){
 
-    int n;
-    int soma = 0;
+    long long int n;
+    long long int soma = 0;
 
     cin >> n;
-    int matriz[n][n];
+    long long int matriz[n][n];
 
     for(int i = 0; i < n; i++ ){
         for(int j = 0; j < n; j++ ){
@@ -19,7 +19,7 @@ int main(){
 
 
     for(int i = 0; i < n; i++){
-            int elemento = matriz[i][n - 1 - i];
+            long long int elemento = matriz[i][n - 1 - i];
             if(elemento % 2 != 0){
                 soma += elemento;
             }
